Added email address and corrected international number
Added surveys.ons.gov.uk as an option if details not as expected, for people who prefer to email rather than use the phone.

Also, the international number had been put in with the leading 0 after the country code, which may have cause a problem for anyone using the link.

Added the nowrap class so phone number remains on one line.