Update helper to be more strict
If any key peice of information is missing from a message, the helper
will raise an error where it used to just put None in the message json
and move on.  This should hopefully help to stop junk data getting into the
system at the slight cost of the user experiance.