Fix secure message usage of requests session (#344)
ras-frontstage uses Session (http://docs.python-requests.org/en/master/user/advanced/) to get retry functionality if a HTTP request fails. When a Session is created the TCP connection is continued to be reused which leads to long running/open TCP connections. Cloundfoundry may do some housekeeping and close long running TCP connections incase they are connection that haven't correctly been closed. If that connection is closed it should reconnect but we're not experiencing this behaviour gracefully when running the services in CF. (https://docs.cloudfoundry.org/services/route-services.html)

All of the above was causing us to experience 500 errors occasionally when sending messages.

Instead of creating one session at the beginning when the application starts we will now create a new session for each request.