[TBL] Call services directly (XL) (#333)
This is in an effort to deprecate frontstage-api by refactoring to use in-house controllers that replicate the implementation from fronstage-api.

This also moves to use the SERVICE_NAME_URL definition for each of the service URLs.

Uses the requests calls directly. ConnectionErrors are handled in the flask error_handler module.